Add changes in functionalities
