+++ index.html
@@ -32,10 +32,10 @@
         </div>
       </nav>
 
-<!--     <nav id="books-section" class="all_books mt-5"> -->
-       <!-- <h1 class="page-title pt-5">All awesome books</h1> -->
+
+     
         <p id="current-time" class="float-end fixed-top mt-5 text-end pt-1"></p>
-<!--     </nav> -->
+
     <section id="books-added" class="px-3">
       <h1 class="page-title">All awesome books</h1>
         <div id="books-container">
